refactor test deploy app
